Move CodableFeedStore to its own file and production code
