+++ Tests/FeedStoreChallengeTests.swift
@@ -5,103 +5,6 @@
 import XCTest
 import FeedStoreChallenge
 
-class CodableFeedStore: FeedStore {
-	
-	private struct Cache: Codable {
-		
-		let items: [CodableFeedImage]
-		let timeStamp: Date
-		
-		var local: [LocalFeedImage] {
-			
-			return items.map { $0.localFeed }
-		}
-	}
-	
-	private struct CodableFeedImage: Codable {
-		
-		private let id: UUID
-		private let description: String?
-		private let location: String?
-		private let imageURL: URL
-		
-		init(_ feed: LocalFeedImage) {
-			self.id = feed.id
-			self.description = feed.description
-			self.location = feed.location
-			self.imageURL = feed.url
-		}
-		
-		var localFeed: LocalFeedImage {
-			return LocalFeedImage(id: id
-				, description: description
-				, location: location
-				, url: imageURL)
-		}
-	}
-	
-	private let storeUrl: URL
-	
-	init(storeUrl: URL) {
-		
-		self.storeUrl = storeUrl
-	}
-	
-	private let queue = DispatchQueue(label: "CodableFeedStore", qos: .userInitiated, attributes: .concurrent)
-	
-	func deleteCachedFeed(completion: @escaping DeletionCompletion) {
-		
-		let storeUrl = self.storeUrl
-		queue.async(flags: .barrier) {
-			
-			guard FileManager.default.fileExists(atPath: storeUrl.path) else {
-				return completion(nil)
-			}
-			
-			do {
-				try FileManager.default.removeItem(at: self.storeUrl)
-				completion(nil)
-			} catch {
-				completion(error)
-			}
-		}
-	}
-	
-	func insert(_ feed: [LocalFeedImage], timestamp: Date, completion: @escaping InsertionCompletion) {
-		
-		let storeUrl = self.storeUrl
-		queue.async(flags: .barrier) {
-			
-			do {
-				
-				let encodedData = try! JSONEncoder().encode(Cache(items: feed.map { CodableFeedImage($0) } , timeStamp: timestamp))
-				try encodedData.write(to: storeUrl)
-				completion(nil)
-			} catch {
-				
-				completion(error)
-			}
-		}
-	}
-	
-	func retrieve(completion: @escaping RetrievalCompletion) {
-		
-		let storeUrl = self.storeUrl
-		queue.async {
-			guard let encodedData = try? Data(contentsOf: storeUrl) else {
-				return completion(.empty)
-			}
-			
-			do {
-				let decodedData = try JSONDecoder().decode(Cache.self, from: encodedData)
-				completion(.found(feed: decodedData.local, timestamp: decodedData.timeStamp))
-			} catch {
-				completion(.failure(error))
-			}
-			
-		}
-	}
-}
 
 class FeedStoreChallengeTests: XCTestCase, FeedStoreSpecs {
